requirements.txt
# Core
django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
djangorestframework-simplejwt==5.3.1
dj-database-url==2.1.0

# Auth & MFA
django-otp==1.5.1
qrcode[pil]==7.4.2
python-ldap==3.4.4
django-auth-ldap==4.6.0

# Storage
boto3==1.34.74               # S3 / compatible object storage
django-storages==1.14.2

# Search
elasticsearch==8.12.1
elasticsearch-dsl==8.9.0
django-elasticsearch-dsl==8.0

# Task queue
celery==5.3.6
redis==5.0.3
django-celery-beat==2.6.0
django-celery-results==2.5.1

# WebSocket / Channels
channels==4.0.0
channels-redis==4.2.0

# Document processing
python-docx==1.1.0
openpyxl==3.1.2
pdfplumber==0.11.0
Pillow==10.3.0
python-magic==0.4.27

# DB
mysqlclient==2.2.4

# Utils
gunicorn==22.0.0
whitenoise==6.6.0
django-auditlog==3.0.0
shortuuid==1.0.13
